pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "parenttext_pipeline"
dynamic = ["version"]
authors = [
    {name = "IDEMS International", email = "contact@idems.international"},
]
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities",
]
dependencies = [
    "beautifulsoup4~=4.12",
    "packaging~=21.3",
    "python-dotenv",
    "rapidpro-abtesting @ git+https://github.com/IDEMSInternational/rapidpro_abtesting.git@master",
    "requests~=2.31",
    "google-cloud-storage~=3.2.0",
    "rpft @ git+https://github.com/IDEMSInternational/rapidpro-flow-toolkit.git@1.13.0",
]

[project.scripts]
rpimport = "parenttext_pipeline.importer:cli"

[tool.setuptools_scm]
version_file = "src/parenttext_pipeline/_version.py"
